Moves kotlin dsl to gateway-core
