--- spring-cloud-gateway-kotlin-extensions/pom.xml
@@ -1,85 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-		 xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-	<modelVersion>4.0.0</modelVersion>
-
-	<parent>
-		<groupId>org.springframework.cloud</groupId>
-		<artifactId>spring-cloud-gateway</artifactId>
-		<version>2.0.0.BUILD-SNAPSHOT</version>
-		<relativePath>..</relativePath> <!-- lookup parent from repository -->
-	</parent>
-	<artifactId>spring-cloud-gateway-kotlin-extensions</artifactId>
-	<packaging>jar</packaging>
-	<name>Spring Cloud Gateway Kotlin Extensions</name>
-	<description>Spring Cloud Gateway Kotlin Extensions</description>
-	<properties>
-		<main.basedir>${basedir}/..</main.basedir>
-		<kotlin.version>1.1.4-3</kotlin.version>
-	</properties>
-
-	<dependencies>
-		<dependency>
-			<groupId>org.springframework.cloud</groupId>
-			<artifactId>spring-cloud-gateway-core</artifactId>
-		</dependency>
-		<dependency>
-			<groupId>org.jetbrains.kotlin</groupId>
-			<artifactId>kotlin-stdlib-jre8</artifactId>
-			<version>${kotlin.version}</version>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-webflux</artifactId>
-			<optional>true</optional>
-		</dependency>
-		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>io.projectreactor</groupId>
-			<artifactId>reactor-test</artifactId>
-			<scope>test</scope>
-		</dependency>
-		<dependency>
-			<groupId>org.assertj</groupId>
-			<artifactId>assertj-core</artifactId>
-			<scope>test</scope>
-		</dependency>
-	</dependencies>
-	
-	<build>
-    	<sourceDirectory>src/main/kotlin</sourceDirectory>
-    	<testSourceDirectory>src/test/kotlin</testSourceDirectory>
-		<plugins>
-			<plugin>
-				<artifactId>kotlin-maven-plugin</artifactId>
-				<configuration>
-					<jvmTarget>1.8</jvmTarget>
-				</configuration>
-				<groupId>org.jetbrains.kotlin</groupId>
-				<version>${kotlin.version}</version>
-
-				<executions>
-					<execution>
-						<id>compile</id>
-						<phase>compile</phase>
-						<goals> 
-							<goal>compile</goal> 
-						</goals>
-					</execution>
-
-					<execution>
-						<id>test-compile</id>
-						<phase>test-compile</phase>
-						<goals> 
-							<goal>test-compile</goal> 
-						</goals>
-					</execution>
-				</executions>
-			</plugin>
-    	</plugins>
-	</build>
-</project>
